config: split the test module by concern
Problem: `src/config/tests.rs` crossed the repo's 1000-line file limit
once the runtime-root-override tests and the `ui.expand_worktrees`
default landed in it, so the pre-push hook refused the branch. The
file had also stopped being navigable: config discovery, per-section
TOML roundtrips, and theme resolution all lived in one flat list of
seventy tests.

Approach: turn it into `src/config/tests/` with the shared `MockEnv`
and `path()` helpers in `mod.rs` and three submodules along the seams
that were already there: `paths` for which config file wins and where
a save lands, `fields` for section defaults, roundtrips, and scan-root
resolution, `theme` for name resolution through `load_with_env`.
Children see the parent's imports through `use super::*`, so no test
body changed. Raising or excluding the limit was the alternative and
it just defers the same split.

Result: 60/366/495/126 lines across the four files, all well under the
limit, with no test added, removed, or edited. `cargo test --bin
gitpane` still reports 443 passing.
